backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend de la prueba tecnica"
authors = [
    {name = "AlexPerez31"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "passlib[bcrypt] (==1.7.4)",
    "bcrypt (==4.0.1)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",   
]

[tool.poetry.dependencies]
httpx = "^0.27.0"
pytest-asyncio = "^0.23.5"


